+++ packages/code-du-travail-frontend/src/api/modules/items/queries.ts
@@ -49,6 +49,7 @@ export const getSearchBySourceSlugBody = ({
       "dismissalProcess",
       "meta_title",
       "type",
+      "content", // pour les contributions génériques et conventionnelles
       "ccSupported", // pour une fiche service-public dans une contribution
       "ccnSlug", // pour une contribution conventionnelle
       "ccnShortTitle", // pour une contribution conventionnelle
