+++ packages/code-du-travail-frontend/src/common/Html.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import xss, { escapeAttrValue } from "xss";
+import { xssWrapper } from "../lib";
 
 type Props = {
   children: string;
@@ -8,38 +8,13 @@ type Props = {
   inline?: boolean;
 };
 
-/**
- * List of tags that are allowed in the HTML
- * @type {string[]}
- * webcomponent-tooltip is used as an overlay for the definition of the words
- */
-const whiteListTags = ["webcomponent-tooltip", "webcomponent-tooltip-cc"];
-
-/**
- * List of attributes that are allowed for the tags scan in the HTML
- * @type {string[]}
- * class is used for modeles-de-courrier
- */
-const whiteListAttr = ["class", "rel", "href", "target"];
-
 const Html = ({ children, as = "div", ...props }: Props): JSX.Element => {
   return (
     <Div
       {...props}
       as={as}
       dangerouslySetInnerHTML={{
-        __html: xss(children, {
-          onIgnoreTag: function (tag, html, _options) {
-            if (whiteListTags.some((whiteTag) => whiteTag === tag)) {
-              return html;
-            }
-          },
-          onTagAttr: function (_tag, name, value, _isWhiteAttr) {
-            if (whiteListAttr.some((whiteAttr) => whiteAttr === name)) {
-              return name + '="' + escapeAttrValue(value) + '"';
-            }
-          },
-        }),
+        __html: xssWrapper(children),
       }}
     />
   );
