+++ packages/code-du-travail-frontend/src/contributions/DisplayContentContribution.tsx
@@ -2,7 +2,7 @@ import { Accordion, theme } from "@socialgouv/cdtn-ui";
 import styled from "styled-components";
 import { FicheServicePublic } from "../fiche-service-public";
 import parse, { domToReact } from "html-react-parser";
-import xss from "xss";
+import { xssWrapper } from "../lib";
 
 export const ContentSP = ({ raw }) => {
   return (
@@ -64,7 +64,7 @@ type Props = {
 const DisplayContentContribution = ({
   content,
 }: Props): string | JSX.Element | JSX.Element[] => {
-  return parse(xss(content), options(3));
+  return parse(xssWrapper(content), options(3));
 };
 
 const { spacings } = theme;
