+++ packages/code-du-travail-frontend/src/lib/index.ts
@@ -4,3 +4,4 @@ export * from "./url";
 export * from "./object";
 export * from "./array";
 export * from "./date";
+export * from "./xss";
