+++ packages/code-du-travail-frontend/src/outils/Components/TextQuestion.tsx
@@ -2,12 +2,12 @@ import { Input, InputDate, theme } from "@socialgouv/cdtn-ui";
 import React, { FunctionComponent, useEffect, useState } from "react";
 
 import styled from "styled-components";
-import xss from "xss";
 import Html from "../../common/Html";
 import { Error } from "../common/ErrorField";
 import { Question, Tooltip } from "../common/Question";
 import { SmallText } from "../common/stepStyles";
 import { SubLabel } from "./SelectQuestion";
+import { xssWrapper } from "../../lib";
 
 type Props = {
   onChange: (value: string) => void;
@@ -87,7 +87,7 @@ export default function TextQuestion({
         <Error>
           <div
             dangerouslySetInnerHTML={{
-              __html: xss(error),
+              __html: xssWrapper(error),
             }}
           />
         </Error>
